+++ Source/Lightning/LightningCore/StreamInterface.cpp
@@ -163,9 +163,9 @@ Rune AsciiEncoding::Read(IStreamClass& stream)
 Integer Utf8Encoding::Write(Rune rune, IStreamClass& stream)
 {
   byte utf8Bytes[4] = {0};
-  size_t encodedByteCount = Plasma::UTF8::UnpackUtf8RuneIntoBuffer(rune.mValue, utf8Bytes);
+  Integer encodedByteCount = static_cast<Integer>(Plasma::UTF8::UnpackUtf8RuneIntoBuffer(rune.mValue, utf8Bytes));
 
-  for (size_t i = 0; i < encodedByteCount; ++i)
+  for (Integer i = 0; i < encodedByteCount; ++i)
   {
     // If we fail to write out a byte, return the amount we've already written
     // (works because of 0 based indexing)
@@ -346,7 +346,7 @@ bool IStreamClass::ValidateArray(ArrayClass<Byte>& data, Integer byteStart, Inte
 
 Integer IStreamClass::Write(ArrayClass<Byte>& data)
 {
-  return this->Write(data, 0, data.NativeArray.Size());
+    return this->Write(data, 0, static_cast<Integer>(data.NativeArray.Size()));
 }
 
 Integer IStreamClass::WriteRune(Rune rune)
